Refactor lang2lcode / treebank_to_short_name
